Move Astro classes to separate sub-package
